Double precisioning all the geometry, improved gamma/diffusion

+++ src/interval.rs
@@ -1,19 +1,19 @@
-use std::f32::{INFINITY, NEG_INFINITY};
+use std::f64::{INFINITY, NEG_INFINITY};
 
 pub struct Interval {
-    pub min: f32,
-    pub max: f32,
+    pub min: f64,
+    pub max: f64,
 }
 
 impl Interval{
-    pub fn new(min: f32, max:  f32) -> Self {
+    pub fn new(min: f64, max:  f64) -> Self {
         Self { min, max }
     }
 
-    pub fn _size(&self) -> f32 { self.max - self.min }
-    pub fn _contains(&self, x: f32) -> bool { self.min <= x && x <= self.max }
-    pub fn surrounds(&self, x: f32) -> bool { self.min < x && x < self.max }
-    pub fn clamp(&self, x: f32) -> f32 {if x < self.min { self.min } else if x > self.max { self.max } else { x } }
+    pub fn _size(&self) -> f64 { self.max - self.min }
+    pub fn _contains(&self, x: f64) -> bool { self.min <= x && x <= self.max }
+    pub fn surrounds(&self, x: f64) -> bool { self.min < x && x < self.max }
+    pub fn clamp(&self, x: f64) -> f64 {if x < self.min { self.min } else if x > self.max { self.max } else { x } }
 }
 
 pub const _EMPTY: Interval = Interval{ min: INFINITY, max: NEG_INFINITY};
